+++ sdk/keyvault/azure_security_keyvault_keys/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- `ResourceExt` canonicalizes URL path segments ([#2177](https://github.com/Azure/azure-sdk-for-rust/issues/2177))
+
 ### Other Changes
 
 ## 0.1.0 (2025-02-18)